add Abstract Widget
