+++ Resources/views/dashboard/dashboard.html.twig
@@ -2,8 +2,6 @@
 {% trans_default_domain 'TkuskaDashboardBundle' %}
 {% block page_head %}{% trans %}dashboard{% endtrans %} {{ renderWidgetSelector() }} {%endblock %}
 
-
-
 {% block head_stylesheets %}
 {{ parent() }}
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/gridstack.js/0.4.0/gridstack.min.css"/>
@@ -23,7 +21,7 @@
     {{ parent() }}
 
     {{ renderWidgetSelector() }}
-    <h1> Dashboard </h1>
+    <h1 class="title"> Dashboard </h1>
     <div class="grid-stack" data-gs-animate="yes">
         {% for widget in widgets %}
             {{ renderWidget(widget) }}
